Fixed Errors in backend

+++ HexaPedal_backend/src/main/java/com/hexpedal/backend/dto/BillingHistoryDto.java
@@ -7,7 +7,7 @@ import java.time.Instant;
 public record BillingHistoryDto(
         Integer rideId,
         Instant startDateTime,
-        Long bikeId,
+        Integer bikeId,
         String originStation,
         String arrivalStation,
         double distance,
